+++ weblate/addons/tasks.py
@@ -5,11 +5,14 @@
 from __future__ import annotations
 
 import os
+from datetime import timedelta
 
 from celery.schedules import crontab
+from django.conf import settings
 from django.db.models import F, Q
 from django.http import HttpRequest
 from django.utils import timezone
+from django.utils.timezone import now
 from lxml import html
 
 from weblate.addons.events import AddonEvent
@@ -125,6 +128,16 @@ def daily_addons(modulo: bool = True) -> None:
     )
 
 
+@app.task(trail=False)
+def cleanup_addon_activity_log() -> None:
+    """Cleanup old add-on activity log entries."""
+    from weblate.addons.models import AddonActivityLog
+
+    AddonActivityLog.objects.filter(
+        created__lt=now() - timedelta(days=settings.ADDON_ACTIVITY_LOG_EXPIRY)
+    ).delete()
+
+
 @app.task(
     trail=False,
     autoretry_for=(WeblateLockTimeoutError,),
@@ -139,3 +152,8 @@ def postconfigure_addon(addon_id: int, addon=None) -> None:
 @app.on_after_finalize.connect
 def setup_periodic_tasks(sender, **kwargs) -> None:
     sender.add_periodic_task(crontab(minute=45), daily_addons.s(), name="daily-addons")
+    sender.add_periodic_task(
+        crontab(hour=0, minute=40),  # Not to run on minute 0 to spread the load
+        cleanup_addon_activity_log.s(),
+        name="cleanup-addon-activity-log",
+    )