+++ docs/changes.rst
@@ -5,10 +5,13 @@ Not yet released.
 
 **New features**
 
+* :ref:`addons` activity log for tracking add-on activity.
+
 **Improvements**
 
 * :ref:`subscriptions` now include strings which need updating.
 * Improved compatibility with password managers.
+* Improved tracking of uploaded changes.
 
 **Bug fixes**